Options for chekbox lists to select: All, None, Invert. This is on by default, but can be turned off for a field in the config file, with 'skip_selector_control'
A (single) "All option" can be specified by setting the checkbox class to checkBoxAllOption.  This will be used instead
of selecting all the individual elements.

For reports, with PICK_MULTI vars, if the first option has code of (ANY|ALL|-1) and label starts with (ALL|ANY) then that
option will be output as an all option.

FIXME: there is stuff in alerts that can and should go.
FIXME: PICK_MULTI should have an explicit specifier for all option.  See comments in code.