For engine_perm, no longer check per-record perms if user has general perms (has_perm())
This was causing performance issues, especially for tables or more so complex
views with high DB costs.  And particularly noticeable on client pages.
This issue went undetected for a long time, largely because super users
were not effected, as their is a hard-coded check for them at the top of the function.